Move Delaunay code to dedicated module
